examples: split the Tier C editor's Win32/X11 code into per-OS TUs
CustomUiWindow.hpp carried both platform implementations of its author
window inline, guarded by a #if/#elif that made the file two editors in a
trenchcoat and dragged <windows.h>/<X11/Xlib.h> (plus their macro-hygiene
#undefs) into the middle of every plug-in TU that includes it.

Hide the window behind a forward-declared pimpl: the header now declares
only the avnd::gui_windowed_ui contract (open/close/idle/size) plus the
message-bus endpoints the bindings introspect, and holds a
std::unique_ptr<struct impl>. The two implementations move to sibling
translation units, CustomUiWindow.win32.cpp (raw Win32 + GDI) and
CustomUiWindow.x11.cpp (raw Xlib), each compiled into the plug-in modules
by examples.cmake for its platform (and linking X11 there, as before). The
header pulls in no windowing header at all -- exactly how a real toolkit
editor would hide behind its own opaque handle.

Behaviour is unchanged: the window class name, 360x120 size, gain gesture
flow and message bus are identical. test_custom_ui_window still passes
(clap module, live X display); vst3/vintage editor modules still build.
